refactor(NameHandler): 입력 문자열을 이름별로 쪼개는 기능을 별도의 클래스로 분리
기존 CarNames에 있던 이름 쪼개기 기능을 NameHandler라는 새로운 클래스로 분리
