+++ src/main/java/racingcar/utils/NameHandler.java
@@ -1,0 +1,12 @@
+package racingcar.utils;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class NameHandler {
+
+    public static List<String> splitNames(String in) {
+        return Arrays.asList(in.split(",", -1));
+    }
+
+}
